requirements.txt
PyQt5
pyserial
opencv-python